chore: separate Accordion components from the single file
